Add session controller.
